pyproject.toml
[tool.black]
line-length = 99
target-version = ["py37", "py38", "py39", "py310"]
exclude = '''
/(
    \.git|.venv
)/
'''

[tool.isort]
profile = "black"
float_to_top = true
skip_glob = ".venv"

[tool.poetry]
name = "python-loki-client"
version = "0.1.0dev0"
description = "Python client for Grafana Loki"
repository = "https://github.com/expobrain/python-loki-client"
authors = ["Daniele Esposti <daniele.esposti@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "grafana_loki_client" }]

[tool.poetry.dependencies]
python = ">=3.7,<4"
httpx = ">=0.23.0"
attrs = ">=20.1.0"
python-dateutil = ">=2.8.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.961"
flake8 = "^4.0.1"
types-python-dateutil = "^2.8.18"
pyupgrade = "^2.37.0"
openapi-python-client = "^0.11.4"
types-certifi = "^2021.10.8"
click = "<9"
toml = "^0.10.2"
types-toml = "^0.10.7"
types-click = "^7.1.8"
flake8-bugbear = "^22.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = ".venv/"
check_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_subclassing_any = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true

[[tool.mypy.overrides]]
module = ["dotenv"]
ignore_missing_imports = true
